% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmplusRtTriangle.R
\name{clmplus.default}
\alias{clmplus.default}
\title{Fit chain-ladder+ to reverse time triangles.}
\usage{
\method{clmplus}{default}(
  RtTriangle,
  hazard.model = NULL,
  gk.fc.model = "a",
  ckj.fc.model = "a",
  gk.order = c(1, 1, 0),
  ckj.order = c(0, 1, 0),
  ...
)
}
\arguments{
\item{RtTriangle}{RtTriangle object to be fitted.}

\item{hazard.model}{hazard model supported from our package, must be provided as a string. The model can be choosen from:
\itemize{
\item{'a': Age model, this is equivalent to the Mack chain-ladder.}
\item{'ac': Age and cohort effects.}
\item{'ap': Age and cohort effects.}
\item{'apc': Age cohort and period effects.}
\item{'cbd': Cairns-Blake-Dowd mortality model (CBD).}
\item{'lc': Lee-Carter parameters: age and age-period interaction effects.}
\item{'m6': CBD with cohorts.}
\item{'m7': CBD m7 extension.}
\item{'m8': CBD m7 extension.}
}}

\item{gk.fc.model}{model to forecast the cohort component for the last accident period. It can be either arima ('a') or linear model ('l'). Disregarded for models that do not have a cohort effect.}

\item{ckj.fc.model}{order of the arima model with drift for the calendar year effect extrapolation. Default to (0,1,0).}

\item{gk.order}{order of the arima model with drift for the accident year effect extrapolation. Default to (1,1,0).}

\item{...}{parameters to be passed to clmplus.}
}
\value{
No return value, called as clmplus method default.
}
\description{
This function allows to fit chain-ladder+ models to cumulative payments run-off triangles.
}
\references{
Hiabu, Munir. “On the relationship between classical chain ladder and granular reserving.” 
Scandinavian Actuarial Journal 2017 (2017): 708 - 729.
}
