% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresidualsclmplusmodel.R
\name{plotresiduals.clmplusmodel}
\alias{plotresiduals.clmplusmodel}
\title{Plot the hazard model residuals}
\usage{
\method{plotresiduals}{clmplusmodel}(clmplusmodel, heat.lim = c(-2.5, 2.5))
}
\arguments{
\item{clmplusmodel}{clmplusmodel object to be plotted.}

\item{heat.lim}{limits in the residuals plot.}
}
\value{
No return value, plots the hazard model residuals in triangular form.
}
\description{
This function allows to plot the hazard model residuals on the triangle payments.
}
\examples{
data(sifa.mtpl)
sifa.mtpl.rtt <- AggregateDataPP(cumulative.payments.triangle=sifa.mtpl)
hz.chl<-clmplus(sifa.mtpl.rtt, 'a')
plotresiduals(hz.chl)

}
\references{
Hiabu, Munir. “On the relationship between classical chain ladder and granular reserving.” 
Scandinavian Actuarial Journal 2017 (2017): 708 - 729.
}
