% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReserve.R
\name{predictReserve}
\alias{predictReserve}
\title{Predict reserve cells}
\usage{
predictReserve(object, ...)
}
\arguments{
\item{object}{A fitted reserving model.}

\item{...}{Additional arguments.}
}
\description{
Generic for reserve-cell predictions.
}
