% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtTriangle.R
\name{RtTriangle}
\alias{RtTriangle}
\title{Reverse time triangles}
\usage{
RtTriangle(cumulative.payments.triangle, k = 1/2)
}
\arguments{
\item{cumulative.payments.triangle}{\code{triangle matrix} or \code{matrix array} object, input triangle of cumulative payments.}

\item{k}{\code{numeric}, individual claims exposure in the cell, also known as lost exposure. It must be in the interval (0,1].}
}
\value{
An object of class \code{RtTriangle}. Lists the following elements:
  \item{cumulative.payments.triangle}{\code{triangle matrix} object, input triangle of cumulative payments.}
  
  \item{occurrance}{\code{matrix array} object, the occurrance derived from the input triangle.}
  
  \item{exposure}{\code{matrix array} object, the exposure derived from the input triangle, under the \code{k} claims arrival assumption.}
  
  \item{incremental.payments.triangle}{\code{triangle matrix} object, incremental payments derived from the input.}
  
  \item{J}{\code{integer}, Run-off triangle dimension.}
  
  \item{diagonal}{ \code{numeric}, cumulative payments last diagonal.}
}
\description{
This function allows to define the class of triangles for reverse time models.
}
\examples{
data(sifa.mtpl)
sifa.mtpl.rtt <- RtTriangle(cumulative.payments.triangle=sifa.mtpl)

}
\references{
Hiabu, Munir. “On the relationship between classical chain ladder and granular reserving.” 
Scandinavian Actuarial Journal 2017 (2017): 708 - 729.
}
