% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amases.gtpl}
\alias{amases.gtpl}
\title{Amases GTPL}
\format{
Run-off triangle with 12 development periods.
}
\usage{
amases.gtpl
}
\description{
Dataset of cumulative paid claims for a small italian company in the line of business: general third party liability.
}
\keyword{datasets}
