% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amases.mod}
\alias{amases.mod}
\title{Amases MOD}
\format{
Run-off triangle with 12 development periods.
}
\usage{
amases.mod
}
\description{
Dataset of cumulative paid claims for a small italian company in the line of business: motor or damage.
}
\references{
Savelli, Nino, and Clemente, Gian Paolo. "Lezioni di matematica attuariale delle assicurazioni danni." EDUCatt-Ente per il diritto allo studio universitario dell'Università Cattolica, 2014
}
\keyword{datasets}
