% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amases.mtpl}
\alias{amases.mtpl}
\title{Amases MTPL}
\format{
Run-off triangle with 12 development periods.
}
\usage{
amases.mtpl
}
\description{
Dataset of cumulative paid claims for a small italian company in the line of business: motor third party liability.
}
\keyword{datasets}
