% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmplusRtTriangle.R
\name{clmplus}
\alias{clmplus}
\title{Fit the reserving model to the triangle in reverse time}
\usage{
clmplus(
  RtTriangle,
  hazard.model,
  xc = NULL,
  iter.max = 10000,
  tolerance.max = 1e-06
)
}
\arguments{
\item{RtTriangle}{RtTriangle object to be fitted}

\item{hazard.model}{hazard model supported from our package, must be provided as a string. The model can be choosen from:
\describe{
       \item{'a': Age model, this is equivalent to the Mack chain-ladder.}
       \item{'ac': Age and cohort effects.}
       \item{'ap': Age and cohort effects.}
       \item{'apc': Age cohort and period effects.}
       \item{'cbd': Cairns-Blake-Dowd mortality model (CBD).}
       \item{'m6': CBD with cohorts.}
       \item{'m7': CBD m7 extension.}
       \item{'m8': CBD m7 extension.}
       }}

\item{xc}{xc constant parameter to be set for the m8 model. Default to NULL.}

\item{tolerance.max}{maximum tolerance of parameters difference for convergence for the Newton-Rhapson algorithm implementation.Ignored for other fitting procedures.}

\item{iter.Max}{maximum number of iterations for the Newton-Rhapson algorithm. Ignored for other fitting procedures.}
}
\description{
This function allows to fit the extended chain-ladder models to the run-off triangles.
}
\examples{
data(sifa.mtpl)
sifa.mtpl.rtt <- RtTriangle(cumulative.payments.triangle=sifa.mtpl)
hz.chl=clmplus(sifa.mtpl.rtt, 'a')

}
\references{
Hiabu, Munir. “On the relationship between classical chain ladder and granular reserving.” 
Scandinavian Actuarial Journal 2017 (2017): 708 - 729.
}
