% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAggregateDataPP.R
\name{plot.AggregateDataPP}
\alias{plot.AggregateDataPP}
\title{Plot the payments behavior}
\usage{
\method{plot}{AggregateDataPP}(x, ...)
}
\arguments{
\item{x}{AggregateDataPP to be plotted.}

\item{...}{Arguments to be passed to plot.}
}
\value{
No return value, plots the run-off triangle cumulative payments and incremental payments.
}
\description{
This function allows to define the behavior of the triangle payments.
}
\examples{
data(sifa.mtpl)
sifa.mtpl.pp <- AggregateDataPP(cumulative.payments.triangle=sifa.mtpl)
plot(sifa.mtpl.pp)

}
\references{
Pittarello, Gabriele, Munir Hiabu, and Andrés M. Villegas. "Replicating and extending chain ladder 
via an age-period-cohort structure on the claim development in a run-off triangle." arXiv preprint arXiv:2301.03858 (2023).
}
