% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRtTriangle.R
\name{plot.RtTriangle}
\alias{plot.RtTriangle}
\title{Plot the payments behavior}
\usage{
\method{plot}{RtTriangle}(obj)
}
\arguments{
\item{obj}{RtTriangle to be plotted.}
}
\description{
This function allows to define the behavior of the triangle payments.
}
\examples{
data(sifa.mtpl)
sifa.mtpl.rtt <- RtTriangle(cumulative.payments.triangle=sifa.mtpl)
plot(sifa.mtpl.rtt)

}
\references{
Hiabu, Munir. “On the relationship between classical chain ladder and granular reserving.” 
Scandinavian Actuarial Journal 2017 (2017): 708 - 729.
}
