% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictclmplusmodel.R
\name{predict.clmplusmodel}
\alias{predict.clmplusmodel}
\title{Predict the Reserve using Chain Ladder Plus Models}
\usage{
\method{predict}{clmplusmodel}(
  clmplusmodel,
  gk.fc.model = "a",
  ckj.fc.model = "a",
  gk.order = c(1, 1, 0),
  ckj.order = c(0, 1, 0),
  forecasting_horizon = NULL,
  ...
)
}
\arguments{
\item{gk.fc.model}{\code{character}, model to forecast the cohort component for the last accident period. It can be either arima ('a') or linear model ('l'). Disregarded for models that do not have a cohort effect.}

\item{ckj.fc.model}{\code{character}, model to forecast the calendar period effect. It can be either arima ('a') or linear model ('l'). Disregarded for models that do not have a period effect.}

\item{gk.order}{\code{integer}, order of the arima model with drift for the accident year effect extrapolation. Default to (1,1,0).}

\item{ckj.order}{\code{integer}, order of the arima model with drift for the calendar year effect extrapolation. Default to (0,1,0).}

\item{forecasting_horizon}{\code{integer}, between 1 and the triangle width. Calendar periods ahead for the predictions. Default predictions are to run-off.}

\item{model}{clmplus model to be plotted.}
}
\value{
predictions.
}
\description{
This function allows to define the behavior of the triangle payments.
}
\references{
Pittarello, Gabriele, Munir Hiabu, and Andrés M. Villegas. "Replicating and extending chain ladder 
via an age-period-cohort structure on the claim development in a run-off triangle." arXiv preprint arXiv:2301.03858 (2023).
}
