% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReserve.R
\name{predictReserve}
\alias{predictReserve}
\title{Predict reserves}
\usage{
predictReserve(object, ...)
}
\arguments{
\item{object}{A fitted model or prediction object.}

\item{...}{Additional arguments passed to a method.}
}
\description{
Re-export of [ReSurv::predictReserve()].
}
