% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sifa.gtpl}
\alias{sifa.gtpl}
\title{Sifa GTPL}
\format{
Run-off triangle with 12 development periods.
}
\usage{
sifa.gtpl
}
\description{
Dataset of cumulative paid claims for a medium italian company in the line of business: general third party liability.
}
\keyword{datasets}
