% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sifa.mod}
\alias{sifa.mod}
\title{Sifa MOD}
\format{
Run-off triangle with 12 development periods.
}
\usage{
sifa.mod
}
\description{
Dataset of cumulative paid claims for a medium italian company in the line of business: general third party liability.
}
\keyword{datasets}
