% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sifa.mtpl}
\alias{sifa.mtpl}
\title{Sifa MTPL}
\format{
Run-off triangle with 12 development periods.
}
\usage{
sifa.mtpl
}
\description{
Dataset of cumulative paid claims for a medium italian company in the line of business: motor third party liability.
}
\keyword{datasets}
